feat: add search input handling, qualifiers, and scroll lock functionality
